Implement event media

+++ src/main/java/com/utem/event_hub_navigation/repo/EventMediaRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.utem.event_hub_navigation.model.EventMedia;
 
-public interface EventMediaRepo extends JpaRepository<EventMedia, Long> {
+public interface EventMediaRepo extends JpaRepository<EventMedia, Integer> {
     // Custom query methods can be defined here if needed
     // For example, to find media by event ID:
